Move server code into the server mixin
