+++ lib/msf/core/exploit/remote/smb/server/share.rb
@@ -6,40 +6,8 @@ module Msf
     # this moment it is capable to share just one file. And the file should
     # live in the root folder "\\".
     module Share
-      include ::Msf::Exploit::Remote::SocketServer
+      include ::Msf::Exploit::Remote::SMB::Server
       include ::Msf::Exploit::Remote::SMB::Server::HashCapture
-
-      module ServiceMixin
-        def start
-          self.listener_thread = Rex::ThreadFactory.spawn('SMBServerListener', false) do
-            print_good 'Running the server'
-            begin
-              run do |server_client|
-                on_client_connect_proc.call(server_client) if on_client_connect_proc
-                true
-              end
-            rescue IOError => e
-              # this 'IOError: stream closed in another thread' is expected, so disregard it
-              wlog("#{e.class}: #{e.message}")
-            end
-          end
-        end
-
-        def thread_local_client
-          connection = @connections.find { |connection| connection.thread.object_id == Thread.current.object_id }
-          connection&.client
-        end
-
-        def stop
-          @socket.close
-        end
-
-        def wait
-          listener_thread.join if listener_thread
-        end
-
-        attr_accessor :listener_thread, :on_client_connect_proc
-      end
 
       # @!attribute share
       #   @return [String] The share portion of the provided UNC.
@@ -72,68 +40,29 @@ module Msf
       end
 
       def start_service(opts = {})
-        @rsock = Rex::Socket::Tcp.create(
-          'LocalHost' => bindhost,
-          'LocalPort' => bindport,
-          'Comm' => _determine_server_comm(bindhost),
-          'Server' => true,
-          'Context' =>
-            {
-              'Msf' => framework,
-              'MsfExploit' => self
-            }
-        )
+        unless opts[:gss_provider]
+          ntlm_provider = Msf::Exploit::Remote::SMB::Server::HashCapture::HashCaptureNTLMProvider.new(
+            allow_anonymous: true,
+            allow_guests: true,
+            listener: self,
+            ntlm_type3_status: nil
+          )
 
-        ntlm_provider = Msf::Exploit::Remote::SMB::Server::HashCapture::HashCaptureNTLMProvider.new(
-          allow_anonymous: true,
-          allow_guests: true,
-          listener: self,
-          ntlm_type3_status: nil
-        )
+          # Set domain name for all future server responses
+          ntlm_provider.dns_domain = datastore['SMBDomain']
+          ntlm_provider.dns_hostname = datastore['SMBDomain']
+          ntlm_provider.netbios_domain = datastore['SMBDomain']
+          ntlm_provider.netbios_hostname = datastore['SMBDomain']
+          opts[:gss_provider] = ntlm_provider
+        end
 
-        # Set domain name for all future server responses
-        ntlm_provider.dns_domain = datastore['SMBDomain']
-        ntlm_provider.dns_hostname = datastore['SMBDomain']
-        ntlm_provider.netbios_domain = datastore['SMBDomain']
-        ntlm_provider.netbios_hostname = datastore['SMBDomain']
-
-        server = RubySMB::Server.new(
-          server_sock: @rsock,
-          gss_provider: ntlm_provider,
-          logger: :stdout
-        )
+        super(opts)
 
         virtual_disk = RubySMB::Server::Share::Provider::VirtualDisk.new(@share)
         virtual_disk.add_dynamic_file("#{@folder_name}#{File::SEPARATOR}#{@file_name}") do
           get_file_contents(client: server.thread_local_client)
         end
         server.add_share(virtual_disk)
-
-        server.extend(ServiceMixin)
-        server.on_client_connect_proc = Proc.new { |client|
-          on_client_connect(client)
-        }
-        self.service = server
-        self.service.start
-
-        print_status("Server is running. Listening on #{bindhost}:#{bindport}")
-      end
-
-      def on_client_connect(client)
-        print_good("Received SMB connection from #{client.peerhost}")
-      end
-
-      def cleanup_service
-        if service
-          begin
-            self.service.stop
-            self.service.wait
-            true
-          rescue ::Exception => e
-            print_error(e.message)
-            false
-          end
-        end
       end
 
       # Setups the server configuration.
